# apps/api/openapi.yaml
openapi: "3.1.0"
info:
  title: ATHENA M&E Platform API
  version: "0.1.0"
  description: |
    Backend API for the ATHENA Monitoring & Evaluation platform.
    All endpoints (except /health and /auth/*) require a valid
    JWT issued via OAuth SSO, carried as an HttpOnly cookie
    (`athena_session`) or an `Authorization: Bearer <token>` header.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your.domain.com
    description: Production

# ── Security scheme ───────────────────────────────────────────
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: athena_session

  # ── Reusable schemas ────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UUID:
      type: string
      format: uuid
    ISOTimestamp:
      type: string
      format: date-time
...
    Entity:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        org_id: { $ref: "#/components/schemas/UUID" }
        entity_type: { type: string }
        external_id: { type: string, nullable: true }
        registered_at: { $ref: "#/components/schemas/ISOTimestamp" }
        metadata: { type: object, additionalProperties: true }
...
    Conflict:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        org_id: { $ref: "#/components/schemas/UUID" }
        entity_id: { $ref: "#/components/schemas/UUID" }
        form_id: { $ref: "#/components/schemas/UUID" }
        canonical_submission_id:
          { $ref: "#/components/schemas/UUID", nullable: true }
        canonical_table: { type: string, nullable: true }
        branch_payload: { type: object, additionalProperties: true }
        resolved: { type: boolean }
        created_at: { $ref: "#/components/schemas/ISOTimestamp" }
...
    Pagination:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        pages: { type: integer }

  # ── Reusable responses ──────────────────────────────────────
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

# ── Default security (override per-endpoint where needed) ─────
security:
  - sessionCookie: []

# ── Paths ─────────────────────────────────────────────────────
paths:
...
  /entities:
    get:
      operationId: listEntities
      summary: Paginated list of entities
      tags: [Entities]
      x-roles: [admin, supervisor]
      parameters:
        - name: entity_type
          in: query
          schema: { type: string }
        - name: registered_after
          in: query
          schema: { type: string, format: date-time }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: Paginated entities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    {
                      type: array,
                      items: { $ref: "#/components/schemas/Entity" },
                    }
                  pagination: { $ref: "#/components/schemas/Pagination" }

    post:
      operationId: createEntity
      summary: Register a new entity
      tags: [Entities]
      x-roles: [admin, supervisor, enumerator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entity_type]
              properties:
                entity_type: { type: string }
                external_id: { type: string }
                metadata: { type: object, additionalProperties: true }
      responses:
        "201":
          description: Entity created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Entity" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /entities/sync:
    get:
      operationId: syncEntities
      summary: Delta sync for Android clients
      tags: [Entities]
      x-roles: [admin, supervisor, enumerator]
      parameters:
        - name: since
          in: query
          required: true
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Entities created or updated after the given timestamp
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Entity" }
  /entities/{id}:
    get:
      operationId: getEntity
      summary: Get a single entity
      tags: [Entities]
      x-roles: [admin, supervisor]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      responses:
        "200":
          description: Entity record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Entity" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── Forms ─────────────────────────────────────────────────────
...
  /conflicts/{id}:
    get:
      operationId: getConflict
      summary: Get canonical and branch payloads side-by-side
      tags: [Quality]
      x-roles: [admin, supervisor]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      responses:
        "200":
          description: Conflict detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conflict" }
        "404": { $ref: "#/components/responses/NotFound" }
  /conflicts/{id}/resolve:
    post:
      operationId: resolveConflict
      summary: Resolve a conflict via merge strategy
      tags: [Quality]
      x-roles: [admin, supervisor]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [merge_strategy]
              properties:
                merge_strategy:
                  type: string
                  enum: [accept_branch, keep_canonical, manual_merge]
                merged_payload:
                  type: object
                  additionalProperties: true
                  description: Required when merge_strategy = manual_merge
      responses:
        "200": { description: Conflict resolved }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── Indicators ────────────────────────────────────────────────
...